Moved files around to clean up project directory
